+++ MagazynKsiazek/MagazynKsiazek/Klasy/Ksiazki.cs
@@ -14,10 +14,10 @@ namespace MagazynKsiazek.Klasy
         public string RokWydania { get; set; }
         public string Gatunek { get; set; }
         public Int32 Ilosc { get; set; }
-        public Double Cena { get; set; }
+        public Decimal Cena { get; set; }
 
         public Ksiazki(int isbn, string tytul, string daneAutora, string wydawnictwo,
-        string rokwydania, string gatunek, Int32 ilosc, Double cena)
+        string rokwydania, string gatunek, Int32 ilosc, Decimal cena)
         {
             ISBN = isbn;
             Tytul = tytul;
